Various small fixes to remove some warnings (conversion etc)
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@7829 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ external/logging/boost/logging/format/formatter/high_precision_time.hpp
@@ -78,8 +78,8 @@ template<class convert = do_convert_format::prepend> struct high_precision_time_
     template<class msg_type> void write_high_precision_time(msg_type & msg, ::boost::posix_time::ptime val) const {
         char_type buffer[64];
 
-        int nanosecs = (int)val.time_of_day().fractional_seconds();
-        int digits = val.time_of_day().num_fractional_digits();
+        int nanosecs = static_cast<int>(val.time_of_day().fractional_seconds());
+        int digits = static_cast<int>(val.time_of_day().num_fractional_digits());
         switch ( digits) {
         case 0: break; // no high precision at all
         case 1: nanosecs *= 100000000; break;
